CLN: #18543, moving periods.pyx
