+++ pandas/_libs/tslibs/period.pyx
@@ -17,7 +17,7 @@ from pandas.compat import PY2
 
 cimport cython
 
-from tslibs.np_datetime cimport (pandas_datetimestruct,
+from pandas._libs.tslibs.np_datetime cimport (pandas_datetimestruct,
                                  dtstruct_to_dt64, dt64_to_dtstruct,
                                  is_leapyear)
 
@@ -25,18 +25,18 @@ from tslibs.np_datetime cimport (pandas_datetimestruct,
 cimport util
 from util cimport is_period_object, is_string_object, INT32_MIN
 
-from missing cimport is_null_datetimelike
+from pandas._libs.missing cimport is_null_datetimelike
 from pandas._libs.tslib import Timestamp
-from tslibs.timezones cimport (
+from pandas._libs.tslibs.timezones cimport (
     is_utc, is_tzlocal, get_utcoffset, get_dst_info)
-from tslibs.timedeltas cimport delta_to_nanoseconds
+from pandas._libs.tslibs.timedeltas cimport delta_to_nanoseconds
 
-from tslibs.parsing import (parse_time_string, NAT_SENTINEL,
-                            _get_rule_month, _MONTH_NUMBERS)
-from tslibs.frequencies cimport get_freq_code
-from tslibs.resolution import resolution, Resolution
-from tslibs.nattype import nat_strings, NaT, iNaT
-from tslibs.nattype cimport _nat_scalar_rules, NPY_NAT
+from pandas._libs.tslibs.parsing import (parse_time_string, NAT_SENTINEL,
+                                         _get_rule_month, _MONTH_NUMBERS)
+from pandas._libs.tslibs.frequencies cimport get_freq_code
+from pandas._libs.tslibs.resolution import resolution, Resolution
+from pandas._libs.tslibs.nattype import nat_strings, NaT, iNaT
+from pandas._libs.tslibs.nattype cimport _nat_scalar_rules, NPY_NAT
 
 from pandas.tseries import offsets
 from pandas.tseries import frequencies
