+++ pandas/compat/pickle_compat.py
@@ -104,7 +104,10 @@ _class_locations_map = {
     ('pandas.tseries.index', 'DatetimeIndex'):
         ('pandas.core.indexes.datetimes', 'DatetimeIndex'),
     ('pandas.tseries.period', 'PeriodIndex'):
-        ('pandas.core.indexes.period', 'PeriodIndex')
+        ('pandas.core.indexes.period', 'PeriodIndex'),
+
+    # 18543 moving period
+    ('pandas._libs.period', 'Period'): ('pandas._libs.tslibs.period', 'Period'),
 }
