+++ pandas/tests/scalar/test_period.py
@@ -10,7 +10,8 @@ import pandas.core.indexes.period as period
 from pandas.compat import text_type, iteritems
 from pandas.compat.numpy import np_datetime64_compat
 
-from pandas._libs import tslib, period as libperiod
+from pandas._libs import tslib
+from pandas._libs.tslibs import period as libperiod
 from pandas._libs.tslibs.parsing import DateParseError
 from pandas import Period, Timestamp, offsets
 from pandas._libs.tslibs.resolution import DAYS, _MONTHS as MONTHS
