+++ pandas/tests/scalar/test_timestamp.py
@@ -16,9 +16,8 @@ from pytz.exceptions import AmbiguousTimeError, NonExistentTimeError
 
 import pandas.util.testing as tm
 from pandas.tseries import offsets, frequencies
-from pandas._libs import period
 from pandas._libs.tslibs.timezones import get_timezone
-from pandas._libs.tslibs import conversion
+from pandas._libs.tslibs import conversion, period
 
 from pandas.compat import long, PY3
 from pandas.util.testing import assert_series_equal
