+++ pandas/tests/test_resample.py
@@ -33,7 +33,7 @@ from pandas.core.resample import (DatetimeIndex, TimeGrouper,
 from pandas.core.indexes.timedeltas import timedelta_range, TimedeltaIndex
 from pandas.util.testing import (assert_series_equal, assert_almost_equal,
                                  assert_frame_equal, assert_index_equal)
-from pandas._libs.period import IncompatibleFrequency
+from pandas._libs.tslibs.period import IncompatibleFrequency
 
 bday = BDay()
